Fixed features and added image compression and base64 encoding
