--- external/gzip/config.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-#ifndef ZLIB_CONST
-#define ZLIB_CONST
-#endif
