--- external/gzip/version.hpp
@@ -1,16 +1,0 @@
-#pragma once
-
-/// The major version number
-#define GZIP_VERSION_MAJOR 1
-
-/// The minor version number
-#define GZIP_VERSION_MINOR 0
-
-/// The patch number
-#define GZIP_VERSION_PATCH 0
-
-/// The complete version number
-#define GZIP_VERSION_CODE (GZIP_VERSION_MAJOR * 10000 + GZIP_VERSION_MINOR * 100 + GZIP_VERSION_PATCH)
-
-/// Version number as string
-#define GZIP_VERSION_STRING "1.0.0"
