+++ src/Commands/LuaCommands.h
@@ -18,8 +18,6 @@ extern "C"
 	#include "../../external/lua/lualib.h"
 }
 
-#pragma comment(lib, "Libs/liblua53.a")
-
 int SDLWindow(lua_State *L);
 
 int moveMouse(lua_State *L);
